docker/docker-wainlux/requirements.txt
Flask==2.3.2
Pillow==11.0.0
pyserial==3.5
# ARMv6/Python 3.11: NumPy 2.x wheel/import is unreliable; pin to known-working branch.
numpy==1.24.3
qrcode==7.4.2
